Update patch set 1
Patch Set 1:

Yes and no. These tests (unit or component tests) are happy now. They are not run (feel free to implement take this ticket then #16399).

I saw that, but too late, this is a perlcritic thing. If desired I will take care of this too. (There may be more perlcritc errors; we don't run the full suit if you remember, but just the files touched (which is why here no errors are reported)

Patch-set: 1
Label: Code-Review=0